requirements.txt
Django==3.2.6
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2-binary